requirements.txt
apache-airflow[postgres,ldap,statsd,ssh,google]==2.2.4
python-dotenv
coverage
invoke
GitPython
flake8
wheel
tailhead
psutil
pyhamcrest
oauth2client==3.0.0
pandas
cryptography
psycopg2-binary
